refactor(vue-next): move pure type definition files directory to src
