+++ packages/hippy-vue-next/src/modules/events.ts
@@ -21,7 +21,7 @@
 import type { ComponentInternalInstance } from '@vue/runtime-core';
 import { callWithAsyncErrorHandling, ErrorCodes } from '@vue/runtime-core';
 
-import type { CallbackType } from '../../types';
+import type { CallbackType } from '../types';
 import type { HippyElement } from '../runtime/element/hippy-element';
 import { lowerFirstLetter } from '../util';
