+++ packages/hippy-vue-next/src/runtime/element/hippy-input-element.ts
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type { NeedToTyped } from '../../../types';
+import type { NeedToTyped } from '../../types';
 import { Native } from '../native';
 
 import { HippyElement } from './hippy-element';
