+++ packages/hippy-vue-next/src/runtime/style/parser.ts
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type { NeedToTyped } from '../../../types';
+import type { NeedToTyped } from '../../types';
 
 
 export interface SelectorType {
