+++ packages/hippy-vue-next/src/types/native-modules.ts
