+++ packages/hippy-vue-next/src/types/native-modules/animation-module.ts
