+++ packages/hippy-vue-next/src/types/native-modules/clip-board-module.ts
