+++ packages/hippy-vue-next/src/types/native-modules/device-event-module.ts
