+++ packages/hippy-vue-next/src/types/native-modules/image-loader-module.ts
