+++ packages/hippy-vue-next/src/types/native-modules/net-info.ts
