+++ packages/hippy-vue-next/src/types/native-modules/network.ts
