+++ packages/hippy-vue-next/src/types/native-modules/test-module.ts
