+++ packages/hippy-vue-next/src/types/native-modules/websocket.ts
