+++ packages/hippy-vue-next/src/types/native-node.ts
