+++ packages/hippy-vue-next/src/util/instance.ts
@@ -19,7 +19,7 @@
  */
 
 import type { ComponentPublicInstance, App } from '@vue/runtime-core';
-import type { NeedToTyped } from '../../types';
+import type { NeedToTyped } from '../types';
 
 /**
  * @public
